requirements.txt
astroid==2.8.5
click==8.0.3
Flask==2.0.2
greenlet==1.1.2
isort==5.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
mysqlclient==2.1.0
platformdirs==2.4.0
pylint==2.11.1
PyMySQL==1.0.2
python-dotenv==0.19.2
SQLAlchemy==1.4.27
toml==0.10.2
typing-extensions==4.0.0
Werkzeug==2.0.2
wrapt==1.13.3
